input worker
